Change to trap ex for local pkgs search
There is some kind of whacky issue in enumerating and running through
the local packages where I think it just needs to skip some folders or
something. The try-catch in this commit will bypass the problem until I
can get back to it and find out what the deal is.